Update main.py
Updated the program to store the account data in a JSON file, instead of a txt file. Also added exception handling when reading and writing from the file so there will be no FileNotFoundErrors. I also added a Search button that will search the data file to see if the user has already stored a password for a website. If so, it shows a message with the saved account information.